Split ItemRef from Retriver
