+++ core/src/main/scala/io/github/scalaquest/core/model/ItemRef.scala
@@ -1,0 +1,3 @@
+package io.github.scalaquest.core.model
+
+trait ItemRef
